compilation fix

+++ Carve/carve/polyline_iter.hpp
@@ -28,12 +28,12 @@ namespace carve {
 
     struct polyline_vertex_iter : public std::iterator<std::random_access_iterator_tag, Vertex *> {
       Polyline *base;
-      ssize_t idx;
+      int64_t idx;
 
       polyline_vertex_iter(Polyline *_base) : base(_base), idx(0) {
       }
 
-      polyline_vertex_iter(Polyline *_base, ssize_t _idx) : base(_base), idx(_idx) {
+      polyline_vertex_iter(Polyline *_base, int64_t _idx) : base(_base), idx(_idx) {
       }
 
       polyline_vertex_iter operator++(int) { return polyline_vertex_iter(base, idx++); }
@@ -52,7 +52,7 @@ namespace carve {
 
 
 
-    static inline ssize_t operator-(const polyline_vertex_iter &a, const polyline_vertex_iter &b) { return a.idx - b.idx; }
+    static inline int64_t operator-(const polyline_vertex_iter &a, const polyline_vertex_iter &b) { return a.idx - b.idx; }
 
     static inline bool operator==(const polyline_vertex_iter&a, const polyline_vertex_iter &b) { return a.idx == b.idx; }
     static inline bool operator!=(const polyline_vertex_iter&a, const polyline_vertex_iter &b) { return a.idx != b.idx; }
@@ -65,12 +65,12 @@ namespace carve {
 
     struct polyline_vertex_const_iter : public std::iterator<std::random_access_iterator_tag, Vertex *> {
       const Polyline *base;
-      ssize_t idx;
+      int64_t idx;
 
       polyline_vertex_const_iter(const Polyline *_base) : base(_base), idx(0) {
       }
 
-      polyline_vertex_const_iter(const Polyline *_base, ssize_t _idx) : base(_base), idx(_idx) {
+      polyline_vertex_const_iter(const Polyline *_base, int64_t _idx) : base(_base), idx(_idx) {
       }
 
       polyline_vertex_const_iter operator++(int) { return polyline_vertex_const_iter(base, idx++); }
@@ -89,7 +89,7 @@ namespace carve {
 
 
 
-    static inline ssize_t operator-(const polyline_vertex_const_iter &a, const polyline_vertex_const_iter &b) { return a.idx - b.idx; }
+    static inline int64_t operator-(const polyline_vertex_const_iter &a, const polyline_vertex_const_iter &b) { return a.idx - b.idx; }
                 
     static inline bool operator==(const polyline_vertex_const_iter&a, const polyline_vertex_const_iter &b) { return a.idx == b.idx; }
     static inline bool operator!=(const polyline_vertex_const_iter&a, const polyline_vertex_const_iter &b) { return a.idx != b.idx; }
@@ -102,25 +102,25 @@ namespace carve {
       return polyline_vertex_const_iter(this, 0);
     }
     inline polyline_vertex_const_iter Polyline::vend() const { 
-      return polyline_vertex_const_iter(this, (ssize_t)vertexCount());
+      return polyline_vertex_const_iter(this, (int64_t)vertexCount());
     }
     inline polyline_vertex_iter Polyline::vbegin() { 
       return polyline_vertex_iter(this, 0);
     }
     inline polyline_vertex_iter Polyline::vend() { 
-      return polyline_vertex_iter(this, (ssize_t)vertexCount());
+      return polyline_vertex_iter(this, (int64_t)vertexCount());
     }
 
 
 
     struct polyline_edge_iter : public std::iterator<std::random_access_iterator_tag, PolylineEdge *> {
       Polyline *base;
-      ssize_t idx;
+      int64_t idx;
 
       polyline_edge_iter(Polyline *_base) : base(_base), idx(0) {
       }
 
-      polyline_edge_iter(Polyline *_base, ssize_t _idx) : base(_base), idx(_idx) {
+      polyline_edge_iter(Polyline *_base, int64_t _idx) : base(_base), idx(_idx) {
       }
 
       polyline_edge_iter operator++(int) { return polyline_edge_iter(base, idx++); }
@@ -139,7 +139,7 @@ namespace carve {
 
 
 
-    static inline ssize_t operator-(const polyline_edge_iter&a, const polyline_edge_iter &b) { return a.idx - b.idx; }
+    static inline int64_t operator-(const polyline_edge_iter&a, const polyline_edge_iter &b) { return a.idx - b.idx; }
                 
     static inline bool operator==(const polyline_edge_iter&a, const polyline_edge_iter &b) { return a.idx == b.idx; }
     static inline bool operator!=(const polyline_edge_iter&a, const polyline_edge_iter &b) { return a.idx != b.idx; }
@@ -152,12 +152,12 @@ namespace carve {
 
     struct polyline_edge_const_iter : public std::iterator<std::random_access_iterator_tag, PolylineEdge *> {
       const Polyline *base;
-      ssize_t idx;
+      int64_t idx;
 
       polyline_edge_const_iter(const Polyline *_base) : base(_base), idx(0) {
       }
 
-      polyline_edge_const_iter(const Polyline *_base, ssize_t _idx) : base(_base), idx(_idx) {
+      polyline_edge_const_iter(const Polyline *_base, int64_t _idx) : base(_base), idx(_idx) {
       }
 
       polyline_edge_const_iter operator++(int) { return polyline_edge_const_iter(base, idx++); }
@@ -176,7 +176,7 @@ namespace carve {
 
 
 
-    static inline ssize_t operator-(const polyline_edge_const_iter&a, const polyline_edge_const_iter &b) { return a.idx - b.idx; }
+    static inline int64_t operator-(const polyline_edge_const_iter&a, const polyline_edge_const_iter &b) { return a.idx - b.idx; }
                 
     static inline bool operator==(const polyline_edge_const_iter&a, const polyline_edge_const_iter &b) { return a.idx == b.idx; }
     static inline bool operator!=(const polyline_edge_const_iter&a, const polyline_edge_const_iter &b) { return a.idx != b.idx; }
@@ -189,13 +189,13 @@ namespace carve {
       return polyline_edge_const_iter(this, 0);
     }
     inline polyline_edge_const_iter Polyline::eend() const { 
-      return polyline_edge_const_iter(this, (ssize_t)edgeCount());
+      return polyline_edge_const_iter(this, (int64_t)edgeCount());
     }
     inline polyline_edge_iter Polyline::ebegin() { 
       return polyline_edge_iter(this, 0);
     }
     inline polyline_edge_iter Polyline::eend() { 
-      return polyline_edge_iter(this, (ssize_t)edgeCount());
+      return polyline_edge_iter(this, (int64_t)edgeCount());
     }
 
   }
